+++ freecad/BulletDesigner/Gui/PreferencesPage.py
@@ -8,7 +8,7 @@ import FreeCAD as App
 import FreeCADGui as Gui
 from PySide import QtCore, QtWidgets
 
-from bullet_designer.Utils.MaterialDatabase import get_material_database
+from freecad.BulletDesigner.Utils.MaterialDatabase import get_material_database
 
 
 class BulletDesignerPreferencesPage: